learn_python_offline.py
This script will download the content from http://learnpythonthehardway.org/book/ website and create offline version of it. When you run the script, the script would ask you to enter an output directory, where it will download the html pages and save them, minus the graphics. It will be a plain html file. The links in the first page would work the same way as it works in the website.